+++ cpp/locality/FilterSANN.h
@@ -21,9 +21,7 @@ public:
      * \param allow_incomplete_shell whether incomplete neighbor shells should
      *                               result in a warning or error
      * */
-    explicit FilterSANN(bool allow_incomplete_shell)
-        : Filter(), m_allow_incomplete_shell(allow_incomplete_shell)
-    {}
+    explicit FilterSANN(bool allow_incomplete_shell) : Filter(allow_incomplete_shell) {}
 
     void compute(const NeighborQuery* nq, const vec3<float>* query_points, unsigned int num_query_points,
                  const NeighborList* nlist, const QueryArgs& qargs) override;
@@ -31,9 +29,6 @@ public:
 private:
     //! warn/raise exception about unfilled shells
     void warnAboutUnfilledNeighborShells(const std::vector<unsigned int>& unfilled_qps) const;
-
-    //! whether incomplete neighbor shell should result in a warning or error
-    bool m_allow_incomplete_shell;
 };
 
 }; };  // namespace freud::locality